requirements.txt
flask
numpy
pandas
scikit-learn
requests
python-dotenv
requests
python-dotenv
matplotlib
seaborn